# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Commerce API Segura",
        "description": "API RESTful para comercio electrónico con Laravel, JWT y Stripe.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor Local de Desarrollo"
        }
    ],
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Registrar un nuevo cliente",
                "operationId": "c7bb21a0d87049e161b8e4c9b0d909f2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Carlos Gómez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "carlos@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario registrado exitosamente"
                    },
                    "422": {
                        "description": "Error de validación en los datos enviados"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Iniciar sesión y obtener Token JWT",
                "operationId": "dd51a484b7fa0f97c7680fdbde14cf5f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "carlos@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login exitoso, devuelve Bearer Token"
                    },
                    "401": {
                        "description": "Credenciales inválidas"
                    }
                }
            }
        },
        "/api/auth/me": {
            "get": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Obtener datos del usuario autenticado",
                "operationId": "b5de204f5e452ef51327c691aa99e77c",
                "responses": {
                    "200": {
                        "description": "Datos del usuario actual"
                    },
                    "401": {
                        "description": "Unauthenticated - Token ausente o inválido"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cerrar sesión e invalidar Token",
                "operationId": "df9bdc6c7a6d9f20b983526b2ec0f698",
                "responses": {
                    "200": {
                        "description": "Sesión cerrada correctamente"
                    },
                    "401": {
                        "description": "Unauthenticated - Token ausente o inválido"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders": {
            "get": {
                "tags": [
                    "Órdenes"
                ],
                "summary": "Listar las órdenes del usuario autenticado",
                "operationId": "955bb77a45ae173a485fadb8befed93f",
                "responses": {
                    "200": {
                        "description": "Lista de órdenes obtenida correctamente"
                    },
                    "401": {
                        "description": "Unauthenticated - Requiere Token JWT"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Órdenes"
                ],
                "summary": "Crear orden y procesar pago",
                "operationId": "2123ba58dadb22bd6f2c8081af562230",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "items",
                                    "payment_method_id"
                                ],
                                "properties": {
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "payment_method_id": {
                                        "type": "string",
                                        "example": "pm_card_visa"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Orden creada y pago procesado con éxito"
                    },
                    "400": {
                        "description": "Error al procesar el pago con Stripe"
                    },
                    "401": {
                        "description": "Unauthenticated - Requiere Token JWT"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/payment/create-session": {
            "post": {
                "tags": [
                    "Pagos"
                ],
                "summary": "Crear una sesión de pago en Stripe",
                "operationId": "c8097a553852bd3a3fb2e6d0e956279e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount",
                                    "product_name"
                                ],
                                "properties": {
                                    "amount": {
                                        "description": "Monto a cobrar",
                                        "type": "number",
                                        "format": "float",
                                        "example": 25.5
                                    },
                                    "product_name": {
                                        "description": "Nombre del producto o servicio",
                                        "type": "string",
                                        "example": "Suscripción Premium"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sesión creada exitosamente"
                    },
                    "401": {
                        "description": "No autorizado - Token ausente o inválido"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "500": {
                        "description": "Error interno de Stripe"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener catálogo de productos",
                "operationId": "d97d1c3727cbac49708b85f7d11f7d58",
                "responses": {
                    "200": {
                        "description": "Lista de productos obtenida correctamente"
                    }
                }
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crear producto",
                "operationId": "785644b7d7dd38af4b51dafd0fbc1b6d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "price",
                                    "stock"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Camiseta Deportiva"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Camiseta de alta calidad para entrenamiento"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 25.99
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 50
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado correctamente"
                    },
                    "422": {
                        "description": "Error de validación en los datos del producto"
                    },
                    "401": {
                        "description": "Unauthenticated - Requiere Token JWT"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener un producto específico",
                "operationId": "c87fb2e4d617e7b36c712cad358acdb8",
                "responses": {
                    "200": {
                        "description": "Producto encontrado exitosamente"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Productos"
                ],
                "summary": "Actualizar un producto existente",
                "operationId": "416e4959a3919547eb609d87e3db9135",
                "responses": {
                    "200": {
                        "description": "Producto actualizado correctamente"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Productos"
                ],
                "summary": "Eliminar un producto",
                "operationId": "b49ce9f3cb168b233245548fb4ef36d3",
                "responses": {
                    "200": {
                        "description": "Producto eliminado correctamente"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingresa tu token JWT con el formato: Bearer {token}",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "tags": [
        {
            "name": "Autenticación",
            "description": "Autenticación"
        },
        {
            "name": "Órdenes",
            "description": "Órdenes"
        },
        {
            "name": "Pagos",
            "description": "Pagos"
        },
        {
            "name": "Productos",
            "description": "Productos"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ]
}